requirements.txt
Pillow==6.1.0
vgio==1.1.0
svgwrite==1.3.1
tabulate==0.8.3
watchdog==0.9.0
